djangoapp/requirements.txt
asgiref==3.8.1
Django>=5.1.3
sqlparse==0.5.2
typing_extensions==4.12.2
psycopg2-binary>=2.9.10
djangorestframework>=3.13.1
pyyaml>=5.1
pygments>=2.7.0
django-filter>=1.0.1
django-guardian>=1.1.1
django-cors-headers>=3.11.0
requests>=2.21.1
numpy
pandas
scikit-learn
load-dotenv